ai2/app/requirements.txt
torch
torchvision
fastapi
uvicorn[standard]
Pillow
pydantic
python-multipart